Add a memory extractor.
It can be usefull to initialize a chunck of memory with predetermined
values from a file or an input stream.

For the moment, the only available function is the string parser. For
example, it can convert a string like "0x05 0xff 0xa0" to an
std::vector<std::byte> {5, 255, 160}.

Later, it will probably be used to initialize the chunck of memory exploited by
the processor. And to directly make it easily execute code comming from
a stream...